Follow up 2: Create base preferences store that `PublicPreferencesStore` extends. Create preference observable generator function to make creation of observable functions easier.
