+++ app/src/main/java/mozilla/lockbox/store/PublicPreferencesStore.kt
@@ -7,10 +7,7 @@
 package mozilla.lockbox.store
 
 import android.content.Context
-import android.content.SharedPreferences
-import android.preference.PreferenceManager
 import io.reactivex.Observable
-import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.rxkotlin.addTo
 import mozilla.lockbox.R
 import mozilla.lockbox.action.ItemListSort
@@ -18,35 +15,22 @@ import mozilla.lockbox.action.SettingsAction
 import mozilla.lockbox.extensions.filterByType
 import mozilla.lockbox.flux.Dispatcher
 
-open class PublicPreferencesStore(
-    val dispatcher: Dispatcher = Dispatcher.shared
-) {
-    internal val compositeDisposable = CompositeDisposable()
-    private lateinit var sharedPrefs: SharedPreferences
+open class PublicPreferencesStore: DefaultPreferencesStore(Dispatcher.shared) {
 
     companion object {
         val shared = PublicPreferencesStore()
     }
 
     private lateinit var itemListSortKey: String
-    val itemListSortObservable: Observable<Int> = Observable.create<Int> { emitter ->
-
-        val sharedPreferencesListener = SharedPreferences.OnSharedPreferenceChangeListener { sharedPreferences, key ->
-            if (key == itemListSortKey) {
-                emitter.onNext(ItemListSort.idFromOrdinal(sharedPreferences.getInt(itemListSortKey,ItemListSort.ALPHABETICALLY.ordinal)))
-            }
+    val itemListSortObservable: Observable<Int> by lazy {
+        val onNext: (String, Int) -> Int = { key, default ->
+            ItemListSort.idFromOrdinal(sharedPrefs.getInt(key,default))
         }
-
-        emitter.setCancellable {
-            sharedPrefs.unregisterOnSharedPreferenceChangeListener(sharedPreferencesListener)
-        }
-
-        emitter.onNext(ItemListSort.idFromOrdinal(sharedPrefs.getInt(itemListSortKey,ItemListSort.ALPHABETICALLY.ordinal)))
-        sharedPrefs.registerOnSharedPreferenceChangeListener(sharedPreferencesListener)
+        createObservableForKey(itemListSortKey, ItemListSort.ALPHABETICALLY.ordinal, onNext)
     }
 
-    fun applyContext(context: Context) {
-        sharedPrefs = PreferenceManager.getDefaultSharedPreferences(context)
+    override fun applyContext(context: Context) {
+        super.applyContext(context)
 
         itemListSortKey = context.resources.getString(R.string.shared_prefs_sort_option_key)
